Refactore the frontend to remove the recursive function. Remove dead code too.
